requirements.txt
pillow
sentence-transformers
